pyproject.toml
[project]
name = "nonebot-plugin-bing-chat"
description = "基于nonebot2的bing-chat软件"
version = "0.6.5"
authors = [
    {name = "Harry", email = "Harry0259@163.com"},
]
readme = "README.md"
license = {text = "GPL-3"}
requires-python = ">=3.8"
dependencies = [
    "nonebot2>=2.0.0b4",
    "nonebot-adapter-onebot>=2.2.0",
    "nonebot_plugin_apscheduler",
    "EdgeGPT>=0.0.59",
]
classifiers=[
        "Framework :: Pydantic",
        "Topic :: Communications",
        "Development Status :: 5 - Production/Stable",
        "License :: OSI Approved :: GNU General Public License (GPL)",
        "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3 :: Only",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Natural Language :: Chinese (Simplified)",
]

[project.urls]
Homepage = "https://github.com/Harry-Jing/nonebot-plugin-bing-chat"


[build-system]
requires = [
  "setuptools >= 67.4.0",
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["nonebot_plugin_bing_chat*"]
